+++ src/models/components/createsupportcopilotconversation.ts
@@ -43,6 +43,10 @@ export type CreateSupportCopilotConversationUserProperties = {
    */
   supportAgentName?: string | null | undefined;
   /**
+   * The email of the support agent assigned to the user. This value is sent by the Inkeep Support Agent Copilot. This value is used to create the graphs on the Inkeep Portal.
+   */
+  supportAgentEmail?: string | null | undefined;
+  /**
    * The type of user. This value is sent by the Inkeep Support Agent Copilot. This value is used to create the graphs on the Inkeep Portal.
    */
   userType?: string | null | undefined;
@@ -115,6 +119,10 @@ export type CreateSupportCopilotConversationMessagesUserProperties = {
    * The name of the support agent assigned to the user. This value is sent by the Inkeep Support Agent Copilot. This value is used to create the graphs on the Inkeep Portal.
    */
   supportAgentName?: string | null | undefined;
+  /**
+   * The email of the support agent assigned to the user. This value is sent by the Inkeep Support Agent Copilot. This value is used to create the graphs on the Inkeep Portal.
+   */
+  supportAgentEmail?: string | null | undefined;
   /**
    * The type of user. This value is sent by the Inkeep Support Agent Copilot. This value is used to create the graphs on the Inkeep Portal.
    */
@@ -268,6 +276,7 @@ export const CreateSupportCopilotConversationUserProperties$inboundSchema:
       identificationType: z.nullable(z.string()).optional(),
       userId: z.nullable(z.union([z.string(), z.number()])).optional(),
       supportAgentName: z.nullable(z.string()).optional(),
+      supportAgentEmail: z.nullable(z.string()).optional(),
       userType: z.nullable(z.string()).optional(),
     }).catchall(z.any()),
     "additionalProperties",
@@ -279,6 +288,7 @@ export type CreateSupportCopilotConversationUserProperties$Outbound = {
   identificationType?: string | null | undefined;
   userId?: string | number | null | undefined;
   supportAgentName?: string | null | undefined;
+  supportAgentEmail?: string | null | undefined;
   userType?: string | null | undefined;
   [additionalProperties: string]: unknown;
 };
@@ -294,6 +304,7 @@ export const CreateSupportCopilotConversationUserProperties$outboundSchema:
     identificationType: z.nullable(z.string()).optional(),
     userId: z.nullable(z.union([z.string(), z.number()])).optional(),
     supportAgentName: z.nullable(z.string()).optional(),
+    supportAgentEmail: z.nullable(z.string()).optional(),
     userType: z.nullable(z.string()).optional(),
     additionalProperties: z.record(z.any()).optional(),
   }).transform((v) => {
@@ -564,6 +575,7 @@ export const CreateSupportCopilotConversationMessagesUserProperties$inboundSchem
       identificationType: z.nullable(z.string()).optional(),
       userId: z.nullable(z.union([z.string(), z.number()])).optional(),
       supportAgentName: z.nullable(z.string()).optional(),
+      supportAgentEmail: z.nullable(z.string()).optional(),
       userType: z.nullable(z.string()).optional(),
     }).catchall(z.any()),
     "additionalProperties",
@@ -575,6 +587,7 @@ export type CreateSupportCopilotConversationMessagesUserProperties$Outbound = {
   identificationType?: string | null | undefined;
   userId?: string | number | null | undefined;
   supportAgentName?: string | null | undefined;
+  supportAgentEmail?: string | null | undefined;
   userType?: string | null | undefined;
   [additionalProperties: string]: unknown;
 };
@@ -590,6 +603,7 @@ export const CreateSupportCopilotConversationMessagesUserProperties$outboundSche
     identificationType: z.nullable(z.string()).optional(),
     userId: z.nullable(z.union([z.string(), z.number()])).optional(),
     supportAgentName: z.nullable(z.string()).optional(),
+    supportAgentEmail: z.nullable(z.string()).optional(),
     userType: z.nullable(z.string()).optional(),
     additionalProperties: z.record(z.any()).optional(),
   }).transform((v) => {