+++ src/models/components/insertmessageevent.ts
@@ -42,6 +42,10 @@ export type InsertMessageEventUserProperties = {
    */
   supportAgentName?: string | null | undefined;
   /**
+   * The email of the support agent assigned to the user. This value is sent by the Inkeep Support Agent Copilot. This value is used to create the graphs on the Inkeep Portal.
+   */
+  supportAgentEmail?: string | null | undefined;
+  /**
    * The type of user. This value is sent by the Inkeep Support Agent Copilot. This value is used to create the graphs on the Inkeep Portal.
    */
   userType?: string | null | undefined;
@@ -146,6 +150,7 @@ export const InsertMessageEventUserProperties$inboundSchema: z.ZodType<
     identificationType: z.nullable(z.string()).optional(),
     userId: z.nullable(z.union([z.string(), z.number()])).optional(),
     supportAgentName: z.nullable(z.string()).optional(),
+    supportAgentEmail: z.nullable(z.string()).optional(),
     userType: z.nullable(z.string()).optional(),
   }).catchall(z.any()),
   "additionalProperties",
@@ -157,6 +162,7 @@ export type InsertMessageEventUserProperties$Outbound = {
   identificationType?: string | null | undefined;
   userId?: string | number | null | undefined;
   supportAgentName?: string | null | undefined;
+  supportAgentEmail?: string | null | undefined;
   userType?: string | null | undefined;
   [additionalProperties: string]: unknown;
 };
@@ -171,6 +177,7 @@ export const InsertMessageEventUserProperties$outboundSchema: z.ZodType<
   identificationType: z.nullable(z.string()).optional(),
   userId: z.nullable(z.union([z.string(), z.number()])).optional(),
   supportAgentName: z.nullable(z.string()).optional(),
+  supportAgentEmail: z.nullable(z.string()).optional(),
   userType: z.nullable(z.string()).optional(),
   additionalProperties: z.record(z.any()).optional(),
 }).transform((v) => {