+++ RELEASES.md
@@ -389,3 +389,13 @@ Based on:
 - [typescript v0.2.4-alpha.34] .
 ### Releases
 - [NPM v0.2.4-alpha.34] https://www.npmjs.com/package/@inkeep/inkeep-analytics/v/0.2.4-alpha.34 - .
+
+## 2025-12-15 20:32:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.676.1 (2.781.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.2.4-alpha.35] .
+### Releases
+- [NPM v0.2.4-alpha.35] https://www.npmjs.com/package/@inkeep/inkeep-analytics/v/0.2.4-alpha.35 - .